+++ src/js/basis.js
@@ -6,6 +6,8 @@ import BasisHamburgerBtn from './_hamburger-btn.js';
 import BasisNavbar from './_navbar.js';
 import BasisPageEffect from './_page-effect.js';
 import BasisSelect from './_select.js';
+import BasisCheckbox from './_checkbox.js';
+import BasisRadio from './_radio.js';
 
 document.addEventListener(
   'DOMContentLoaded',
@@ -16,6 +18,8 @@ document.addEventListener(
     new BasisNavbar();
     new BasisPageEffect();
     new BasisSelect();
+    new BasisCheckbox();
+    new BasisRadio();
   },
   false
 );